Revert "fix a likely bug in exclamation boxes where high behavior ids don't work"
This reverts commit c27b9c0e87b75560ba055d11c01a17442c686b8e.

+++ src/game/behaviors/exclamation_box.inc.c
@@ -104,7 +104,7 @@ static s32 exclamation_replace_model(struct MarioState* m, s32 model) {
     }
 }
 
-void exclamation_box_spawn_contents(struct ExclamationBoxContent *content, s16 itemId) {
+void exclamation_box_spawn_contents(struct ExclamationBoxContent *content, u8 itemId) {
     if (content == NULL) { return; }
     struct MarioState* marioState = nearest_mario_state_to_object(o);
     struct Object* player = marioState ? marioState->marioObj : NULL;
